Fix terminal warnings and remove duplicate terminal.rs file
