+++ kernel/src/lang.rs
@@ -1,11 +1,11 @@
 //! Lang items
 
-use core::fmt::{self, Write};
-use spin::RwLock;
-use drivers::vga::VgaWriter;
-use terminal::{self, TerminalOutput, Stdout};
-use color::{Color, ColorPair};
 use ::halt;
+use color::{Color, ColorPair};
+use core::fmt::{self, Write};
+use drivers::vga::VgaWriter;
+use spin::RwLock;
+use terminal::{self, Stdout, TerminalOutput};
 
 #[lang = "eh_personality"]
 #[no_mangle]
@@ -24,7 +24,7 @@ extern fn panic_fmt(args: fmt::Arguments, file: &'static str, line: u32) -> ! {
 
     let _ = writer.set_cursor_pos(terminal::Point::new(0, 5)); // TODO no
     let _ = writer.set_color(ColorPair::new(Color::Red, Color::Black));
-    let _ = write!(&mut writer, "Panicked at \"{}\", {file}:{line}\n", args, file=file, line=line);
+    let _ = write!(&mut writer, "Panicked at \"{}\", {file}:{line}\n", args, file = file, line = line);
 
-    unsafe { halt() }
+    halt()
 }
